create a basic template for an electron desktop app.
the electron app should have a login screen where a user enter  database credentials. users should be able to save particular profiles (combinations of host, port, database, user, password) as named profiles and switch the active connection easily. when users are logged in, they should be taken to a main page that says "Welcome, ___" where ___ is the result of running "SELECT USER_NAME()" on the active connection.

a local sqlite file will be used to save/reload named database configurations.